+++ app/ui/tab_db.R
@@ -17,5 +17,8 @@ tabPanel("Database", value = "db",
                           dataTableOutput("input_files_table")),
          
          conditionalPanel(condition = "output.files_imported & output.no_errors",
-                          actionButton("add_to_db", ""))
+                          actionButton("add_to_db", "")),
+         
+         conditionalPanel(condition = "output.files_imported & !output.no_errors",
+                          h4("Errors found in input; at least one file is either already in the database or has an error."))
 )