+++ app/ui/tab_main.R
@@ -1,7 +1,11 @@
 # QS trend analysis - ui sidebar
 
 tabPanel("Introduction", value = "main",
-         htmlOutput("main_text_1"),
+         htmlOutput("main_text"),
          
-         htmlOutput("paths")
+         htmlOutput("paths"), 
+         
+         conditionalPanel(condition = "!output.qc_present",
+                          htmlOutput("qc_text"),
+                          dataTableOutput("qc_example_table"))
 )